requirements.txt
torch==1.1
overrides
h5py
boto3
pytorch-pretrained-bert==0.6.2
numpy
scipy

# Data processing
spacy
ftfy
nltk
conllu
pyyaml
pycorenlp==0.3
penman==0.6.2
editdistance
word2number
bs4
lxml
networkx

# Write logs for training visualisation with the Tensorboard application
# Install the Tensorboard application separately (part of tensorflow) to view them.
tensorboardX
